Fixed some minor errors (mainly unsigned to signed).
git-svn-id: https://svn.physiomeproject.org/svn/cmiss/cmgui/trunk@1590 3c1deb5b-d424-0410-962d-aba41a686d42

+++ source/general/postscript.c
@@ -1558,7 +1558,7 @@ the drawable (not sure which)
 						(void)fprintf(postscript.file,"%% PutImage\n");
 						(void)fprintf(postscript.file,"gsave\n");
 						(void)fprintf(postscript.file,"/scan_line %d string def\n",
-							scan_line_length);
+							(int)scan_line_length);
 						(void)fprintf(postscript.file,"0 setgray\n");
 						(void)fprintf(postscript.file,"%d %d translate\n",dest_x,dest_y);
 						(void)fprintf(postscript.file,"%u %u\n",width,height);
@@ -1657,7 +1657,7 @@ the drawable (not sure which)
 						(void)fprintf(postscript.file,"%% PutImage\n");
 						(void)fprintf(postscript.file,"gsave\n");
 						(void)fprintf(postscript.file,"/scan_line %d string def\n",
-							scan_line_length);
+							(int)scan_line_length);
 						(void)fprintf(postscript.file,"%d %d translate\n",dest_x,dest_y);
 						(void)fprintf(postscript.file,"%u %u\n",width,height);
 						(void)fprintf(postscript.file,"8\n");
